Convert guided setup modal to new style
#3078
