+++ static/css/inbox.less
@@ -1231,17 +1231,6 @@ a.fa:hover {
       width: 3em;
     }
   }
-  .modal-footer {
-    text-align: left;
-    .right {
-      float: right;
-    }
-    .error {
-      clear: both;
-      font-weight: bold;
-      color: @error-color;
-    }
-  }
 }
 
 .section {
