--- static/js/modules/guided-setup.js
@@ -1,146 +1,0 @@
-var libphonenumber = require('libphonenumber/utils'),
-    countries = require('../modules/countries');
-
-(function () {
-
-  'use strict';
-
-  var selectOption = function(e) {
-    e.preventDefault();
-    var elem = $(this);
-    elem.closest('.horizontal-options')
-      .find('.selected')
-      .removeClass('selected');
-    elem.addClass('selected');
-    var panel = elem.closest('.panel');
-    var label = [];
-    panel.find('.horizontal-options .selected').each(function() {
-      label.push($(this).text().trim());
-    });
-    panel
-      .addClass('panel-complete')
-      .find('.panel-heading .value')
-      .text(label.join(', '));
-  };
-
-  var updateNumbers = function() {
-    var gatewayNumber = $('#guided-setup [name=gateway-number]').val();
-    var defaultCountryCode = $('#guided-setup [name=default-country-code]').val();
-    var parts = [];
-    if (defaultCountryCode) {
-      parts.push('+' + defaultCountryCode);
-    }
-    if (gatewayNumber) {
-      parts.push(gatewayNumber);
-    }
-    $(this).closest('.panel').find('.panel-heading .value').text(parts.join(', '));
-    if (gatewayNumber && defaultCountryCode) {
-      $(this).closest('.panel').addClass('panel-complete');
-    }
-  };
-
-  var validate = function(translateFilter) {
-    var countryCode = $('#guided-setup [name=default-country-code]').val();
-    var gatewayNumber = $('#guided-setup [name=gateway-number]').val();
-    if (gatewayNumber &&
-        !libphonenumber.validate({ default_country_code: countryCode }, gatewayNumber)) {
-      return {
-        valid: false,
-        error: translateFilter('Phone number not valid')
-      };
-    }
-    return { valid: true };
-  };
-
-  var save = function(UpdateSettings, translateFilter, e) {
-    e.preventDefault();
-
-    var valid = validate(translateFilter);
-    if (!valid.valid) {
-      $('#guided-setup .error').text(valid.error).show();
-      return;
-    }
-
-    $('#setup-wizard-save').addClass('disabled');
-    $('#guided-setup .loader').show();
-    $('#guided-setup .error').hide();
-    var settings = {};
-    var val;
-
-    val = $('#guided-setup [name=gateway-number]').val();
-    if (val) {
-      settings.gateway_number = val;
-    }
-    val = $('#guided-setup [name=default-country-code]').val();
-    if (val) {
-      settings.default_country_code = val;
-    }
-    val = $('#primary-contact-content .horizontal-options .selected').attr('data-value');
-    if (val) {
-      settings.care_coordinator = val;
-    }
-    val = $('#language-preference-content .locale .selected').attr('data-value');
-    if (val) {
-      settings.locale = val;
-    }
-    val = $('#language-preference-content .locale-outgoing .selected').attr('data-value');
-    if (val) {
-      settings.locale_outgoing = val;
-    }
-    val = $('#registration-form-content .horizontal-options .selected').attr('data-value');
-    if (val) {
-      settings.anc_registration_lmp = val === 'true';
-    }
-    val = $('#anonymous-statistics-content .horizontal-options .selected').attr('data-value');
-    if (val) {
-      settings.statistics_submission = val;
-    }
-    UpdateSettings(settings)
-      .then(function() {
-        $('#setup-wizard-save').removeClass('disabled');
-        $('#guided-setup .loader').hide();
-        $('#guided-setup').modal('hide');
-      })
-      .catch(function(err) {
-        $('#setup-wizard-save').removeClass('disabled');
-        $('#guided-setup .loader').hide();
-        console.error('Error updating settings', err);
-        $('#guided-setup .error')
-          .text(translateFilter('Error saving settings'))
-          .show();
-      });
-  };
-
-  var bindSettings = function(Settings) {
-    Settings()
-      .then(function(res) {
-        // don't show any default values - encourage the user to select them
-        if (res.setup_complete) {
-          window.setTimeout(function() {
-            $('#guided-setup [name=default-country-code]').val(res.default_country_code).change();
-            $('#guided-setup [name=gateway-number]').val(res.gateway_number).trigger('input');
-            $('#primary-contact-content a[data-value=' + res.care_coordinator + ']').trigger('click');
-            $('#language-preference-content .locale a[data-value=' + res.locale + ']').trigger('click');
-            $('#language-preference-content .locale-outgoing a[data-value=' + res.locale_outgoing + ']').trigger('click');
-            $('#registration-form-content a[data-value=' + res.anc_registration_lmp + ']').trigger('click');
-            $('#anonymous-statistics-content a[data-value=' + res.statistics_submission + ']').trigger('click');
-          }, 1);
-        }
-      })
-      .catch(function(err) {
-        console.error('Error fetching settings', err);
-      });
-  };
-
-  exports.init = function(Settings, UpdateSettings, translateFilter) {
-    $('#guided-setup').on('click', '.horizontal-options a', selectOption);
-    $('#guided-setup [name=gateway-number]').on('input', updateNumbers);
-    $('#guided-setup [name=default-country-code]').select2({ width: '20em', data: countries.list });
-    $('#guided-setup [name=default-country-code]').on('change', updateNumbers);
-    $('#setup-wizard-save').on('click', function(e) {
-      save(UpdateSettings, translateFilter, e);
-    });
-    bindSettings(Settings);
-  };
-
-}());
