+++ templates/inbox.html
@@ -25,8 +25,6 @@
       <span class="snackbar-content"></span>
     </div>
 
-    <ng-include src="'templates/modals/guided_setup.html'" onload="setupGuidedSetup()"></ng-include>
-
     <script src="<%= baseURL %>/static/dist/inbox.js" type="text/javascript"></script>
     <script src="<%= baseURL %>/static/dist/templates.js" type="text/javascript"></script>
