+++ templates/partials/header.html
@@ -49,7 +49,7 @@
             </a>
           </li>
           <li role="presentation" mm-auth="can_configure">
-            <a role="menuitem" tabindex="-1" rel="external" data-toggle="modal" data-target="#guided-setup">
+            <a role="menuitem" tabindex="-1" ng-click="openGuidedSetup()">
               <i class="fa fa-fw fa-list-ol"></i>
               <span translate>Easy Setup Wizard</span>
             </a>
